refactor(shadcn): quality pass - consolidate mocks, extract AnnotationBadges
Quality improvements per V2 guiding principles:
- Add 'active' badge variant for blue status indicators
- Refactor OAuthDebuggerModal to use proper badge variants
- Extract AnnotationBadges component from Tools/Resources
- Consolidate 16 scattered mock data definitions into /mocks/ directory
- Export types from centralized location

Net reduction: ~500 lines of duplicated/scattered code
